Only one multicomparison widget file
